requirements.txt
sentence-transformer
faiss-cp
numpy
panda
fastapi
uvicorn
scikit-lear
transformers
torc
rapidfuzz
